cleanup

+++ include/display/octo_display.h
@@ -15,7 +15,7 @@ namespace display {
 
 template <size_t STRIP_COUNT, size_t STRIP_LENGTH>
 class OctoDisplay : public Display {
-  static constexpr int total_pixel_count{STRIP_COUNT * STRIP_LENGTH};
+  static constexpr size_t total_pixel_count{STRIP_COUNT * STRIP_LENGTH};
 
   // These buffers need to be large enough for all the pixels.
   // The total number of pixels is "ledsPerStrip * numPins".
@@ -51,7 +51,7 @@ class OctoDisplay : public Display {
   }
 
   virtual void Show() override {
-    for (int i = 0; i < total_pixel_count; ++i) {
+    for (size_t i = 0; i < total_pixel_count; ++i) {
       octo.setPixel(i, pixels[i].r, pixels[i].g, pixels[i].b);
     }
     octo.show();
